Minor Function Changes and Cleanup
The substantive changes in this commit come from the GPUBullyImpactBruteforcer,
which had incorrect numbers of parameters in a couple of function calls.
These have been corrected to match the expected function signatures.

Other changes include additional prinouts/improvements to variable names,
and the removal of the non-BITFS project from the solution file, which
was erroneously included and has led to confusion.